简化动态列表，增加 provider 依赖，未使用
