+++ lib/page/dynamic_page.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-import 'package:gsy_github_app_flutter/bloc/base/base_bloc.dart';
 import 'package:gsy_github_app_flutter/bloc/dynamic_bloc.dart';
 import 'package:gsy_github_app_flutter/common/dao/repos_dao.dart';
 import 'package:gsy_github_app_flutter/common/model/Event.dart';
 import 'package:gsy_github_app_flutter/common/redux/gsy_state.dart';
 import 'package:gsy_github_app_flutter/common/utils/event_utils.dart';
 import 'package:gsy_github_app_flutter/widget/event_item.dart';
-import 'package:gsy_github_app_flutter/widget/state/gsy_bloc_list_state.dart';
 import 'package:gsy_github_app_flutter/widget/pull/gsy_pull_new_load_widget.dart';
 import 'package:redux/redux.dart';
 
@@ -21,78 +19,34 @@ class DynamicPage extends StatefulWidget {
   _DynamicPageState createState() => _DynamicPageState();
 }
 
-class _DynamicPageState extends State<DynamicPage>
-    with
-        AutomaticKeepAliveClientMixin<DynamicPage>,
-        GSYListState<DynamicPage>,
-        WidgetsBindingObserver {
+class _DynamicPageState extends State<DynamicPage> with AutomaticKeepAliveClientMixin<DynamicPage>, WidgetsBindingObserver {
   final DynamicBloc dynamicBloc = new DynamicBloc();
 
   ///控制列表滚动和监听
   final ScrollController scrollController = new ScrollController();
 
+  final GlobalKey<RefreshIndicatorState> refreshIndicatorKey = new GlobalKey<RefreshIndicatorState>();
+
+
   /// 模拟IOS下拉显示刷新
-  @override
   showRefreshLoading() {
     ///直接触发下拉
     new Future.delayed(const Duration(milliseconds: 500), () {
-      scrollController.animateTo(-141,
-          duration: Duration(milliseconds: 600), curve: Curves.linear);
+      scrollController.animateTo(-141, duration: Duration(milliseconds: 600), curve: Curves.linear);
       return true;
     });
   }
 
-  @override
-  bool get wantKeepAlive => true;
 
-  @override
-  requestRefresh() async {
-    await Future.delayed(Duration(seconds: 1));
+  ///下拉刷新数据
+  Future<void> requestRefresh() async {
+    //await Future.delayed(Duration(seconds: 1));
     return await dynamicBloc.requestRefresh(_getStore().state.userInfo?.login);
   }
 
-  @override
-  requestLoadMore() async {
+  ///上拉更多请求数据
+  Future<void> requestLoadMore() async {
     return await dynamicBloc.requestLoadMore(_getStore().state.userInfo?.login);
-  }
-
-  @override
-  bool get isRefreshFirst => false;
-
-  @override
-  BlocListBase get bloc => dynamicBloc;
-
-  @override
-  void initState() {
-    super.initState();
-    ///监听生命周期，主要判断页面 resumed 的时候触发刷新
-    WidgetsBinding.instance.addObserver(this);
-    ///获取网络端新版信息
-    ReposDao.getNewsVersion(context, false);
-  }
-
-  @override
-  void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
-    super.dispose();
-  }
-
-  @override
-  void didChangeDependencies() {
-    if (bloc.getDataLength() == 0) {
-      showRefreshLoading();
-    }
-    super.didChangeDependencies();
-  }
-
-  ///监听生命周期，主要判断页面 resumed 的时候触发刷新
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    if (state == AppLifecycleState.resumed) {
-      if (bloc.getDataLength() != 0) {
-        showRefreshLoading();
-      }
-    }
   }
 
   _renderEventItem(Event e) {
@@ -110,27 +64,65 @@ class _DynamicPageState extends State<DynamicPage>
   }
 
   @override
+  void initState() {
+    super.initState();
+
+    ///请求更新
+    if (dynamicBloc.getDataLength() == 0) {
+      dynamicBloc.changeNeedHeaderStatus(false);
+      showRefreshLoading();
+    }
+
+    ///监听生命周期，主要判断页面 resumed 的时候触发刷新
+    WidgetsBinding.instance.addObserver(this);
+
+    ///获取网络端新版信息
+    ReposDao.getNewsVersion(context, false);
+  }
+
+
+  @override
+  void didChangeDependencies() {
+    if (dynamicBloc.getDataLength() == 0) {
+      showRefreshLoading();
+    }
+    super.didChangeDependencies();
+  }
+
+  ///监听生命周期，主要判断页面 resumed 的时候触发刷新
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (state == AppLifecycleState.resumed) {
+      if (dynamicBloc.getDataLength() != 0) {
+        showRefreshLoading();
+      }
+    }
+  }
+
+  @override
+  bool get wantKeepAlive => true;
+
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    dynamicBloc.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     super.build(context); // See AutomaticKeepAliveClientMixin.
-    return new StoreBuilder<GSYState>(
-      builder: (context, store) {
-        ///BlocProvider 用于管理 bloc 共享，如果不需要共享可以不用
-        ///直接 StreamBuilder 配合即可
-        return BlocProvider<DynamicBloc>(
-          bloc: dynamicBloc,
-          child: GSYPullLoadWidget(
-            bloc.pullLoadWidgetControl,
-            (BuildContext context, int index) =>
-                _renderEventItem(bloc.dataList[index]),
-            requestRefresh,
-            requestLoadMore,
-            refreshKey: refreshIndicatorKey,
-            scrollController: scrollController,
-            ///使用ios模式的下拉刷新
-            userIos: true,
-          ),
-        );
-      },
-    );
+    return GSYPullLoadWidget(
+        dynamicBloc.pullLoadWidgetControl,
+        (BuildContext context, int index) => _renderEventItem(dynamicBloc.dataList[index]),
+        requestRefresh,
+        requestLoadMore,
+        refreshKey: refreshIndicatorKey,
+        scrollController: scrollController,
+
+        ///使用ios模式的下拉刷新
+        userIos: true,
+      );
   }
 }
